Generalize FT example to keep it unbounded from IO and STM.
It's possible to introduce classes similar to the ones used
in free language and have generalized implemntation in the
FT case.
The core idea is to introduce a simple class type and then
you can introduce instance for AppM. Then you can control
the implementation by passing differrent wrappers instead
of AppM. It's possible to generalize even more, but in general
it's coutner-productive.